\name{calib-class}
\docType{class}
\alias{calib-class}
\alias{plot,calib,missing-method}
\alias{print,calib-method}
\alias{show,calib-method}
\alias{summary,calib-method}

\title{Object of class calib}
\description{This is an object generated by using the \code{calib} function on an object of class \link{calib.fit}. 
Contains results from performing calibration.
}
%\section{Objects from the Class}{
%Objects can be created by calls of the form \code{new("calib", ...)}.
%	 ~~ describe objects here ~~ 
%}
\section{Slots}{
	 \describe{
    \item{\code{Estimated.x}:}{Predicted values of x}
    \item{\code{PredStdErr}:}{Predicted standard errors}
    \item{\code{inver.low}:}{Estimated lower confidence intervals using inverse method}
    \item{\code{inver.up}:}{Estimated upper confidence intervals using inverse method}
    \item{\code{wald.low}:}{Esimated lower confidence interval using Wald intervals}
    \item{\code{wald.up}:}{Estimated upper confidence interval using Wald intervals}
    \item{\code{avg.response}:}{y0 values}
    \item{\code{dilution}:}{Dilution factors}
    \item{\code{oor}:}{Values which fall out of range of assay}
    \item{\code{lmdc}:}{Internal use}
    \item{\code{row.names}:}{Internal use}
    \item{\code{labels}:}{Internal use}
    \item{\code{max.x}:}{Internal use}
    \item{\code{extrap}:}{Should values outside the concentration range be estimated}
    \item{\code{repeq}:}{Internal use}
    \item{\code{rname}:}{Internal use}
    \item{\code{conf.level}:}{Values to be used for estimating confidence intervals of
    esimated concentrations}
    \item{\code{type}:}{Model type used (fpl, thpl or lin)}
    \item{\code{mdc}:}{Minimum detectable concentration values}
    \item{\code{truth}:}{If true concentration values are provided 
    in the calib function}
    \item{\code{times}:}{Internal use}
    \item{\code{samp.names}:}{Names of samples}
  }
}
\section{Methods}{
  \describe{
    \item{plot}{Plot results from a calib object.}
    \item{print}{Print results from a calib object.}
    \item{show}{Same as print.}
    \item{summary}{Same as print.}
	 }
}
\author{Perry Haaland, Elaine McVey, Daniel Samarov}
\seealso{
\code{\link{calib.fit-class}, \link{calib.fit}, \link{calib}, \link{plot}, \link{show}, \link{summary},
\link{print}}
}
\examples{
showClass("calib")
}
\keyword{classes}
