\name{coef-methods}
\docType{methods}
\alias{coef-methods}
\alias{coef,ANY-method}
\title{Method coef for `calib' package}
\description{
Returns coefficient information from a \code{calib.fit} object.
}
\section{Methods}{
\describe{

\item{object = "ANY"}{Objects \code{coef} applies to other than \code{calib.fit} objects.}

\item{object = "calib.fit"}{\code{calib.fit} object. }
}}
\keyword{methods}
