\name{coefficients-methods}
\docType{methods}
\alias{coefficients}

\title{Coefficients Method}
\description{
 Extracts the value of the coefficients from an object of class calib.fit
}
\section{Methods}{
\describe{

\item{object = "ANY"}{Extract coefficient value}

\item{object = "calib.fit"}{Extract coefficient values from a calib.fit object}
}}
\seealso{\code{\link{calib.fit}}}}

\keyword{methods}

