\name{experimentData}
\alias{experimentData}
\docType{data}
\title{Experimental Dataset}
\description{
Dataset similar to ELISA dataset
}
\usage{data(experimentData)}
\format{
  A data frame with 72 observations on the following 3 variables.
  \describe{
    \item{\code{group}}{Unique identifier}
    \item{\code{OD}}{A numeric vector of respnses}
    \item{\code{dilution}}{A numeric vector of dilutions}
  }
}

\examples{
data(experimentData)
}
\keyword{datasets}
