\name{fitted-methods}
\docType{methods}
\alias{fitted}
\title{Fitted Method}
\description{
 Extract fitted values from an object of class calib.fit
}
\section{Methods}{
\describe{

\item{object = "ANY"}{Extract fitted values}

\item{object = "calib.fit"}{Extract fitted values from an object of class calib.fit}
}}
\seealso{\code{\link{calib.fit}}}}

\keyword{methods}

