\name{HPLC}
\alias{HPLC}
\docType{data}
\title{High Performance Liquid Chromatography (HPLC) dataset}
\description{
A data set with the results from an HPLC experiment
}
\usage{data(HPLC)}
\format{
  A data frame with 51 observations on the following 2 variables.
  \describe{
    \item{\code{Concentration}}{Concentration of analyte}
    \item{\code{Response}}{Response of assay}
  }
}

\examples{
data(HPLC)

}
\keyword{datasets}
