\name{loq-methods}
\docType{methods}
\alias{loq}

\title{Limit of Quantization (LOQ)}
\description{
 Extracts LOQ value from an object of class calib.fit
}
\section{Methods}{
\describe{

\item{object = "calib.fit"}{Extracts LOQ value}
}}
\seealso{\code{\link{calib.fit}}}}

\keyword{methods}

