\name{mdc-methods}
\docType{methods}
\alias{mdc}

\title{Minimum Detectable Concentration (MDC)}
\description{
 Extracts MDC value from an object of class calib.fit
}
\section{Methods}{
\describe{

\item{object = "calib.fit"}{Extracts MDC value}
}}
\seealso{\code{\link{calib.fit}}}}

\keyword{methods}

