\name{plot-methods}
\docType{methods}
\alias{plot-methods}
\alias{plot,ANY,ANY-method}
\title{Plot method}
\description{
Plotting method for \code{\link{calib.fit}} and \code{\link{calib}}Provides plotting methods 
}
\section{Methods}{
\describe{

\item{x = "ANY", y = "ANY"}{ ~~describe this method here }

\item{x = "calib", y = "missing"}{ ~~describe this method here }

\item{x = "calib.fit", y = "missing"}{ ~~describe this method here }
}}
\keyword{methods}

