\name{print-methods}
\docType{methods}
\alias{print-methods}
\alias{print,ANY-method}
\title{Method print for calib.fit object.}
\description{Prints output relevant to a \code{calib.fit} or \code{calib} object.}
\section{Methods}{
\describe{

\item{x = "ANY"}{Objects other than \code{calib.fit} and \code{calib} for which \code{print} can be used.}

\item{x = "calib"}{\code{calib} object.}

\item{x = "calib.fit"}{\code{calib.fit} object.}
}}
\keyword{methods}
