\name{resid-methods}
\docType{methods}
\alias{resid-methods}
\alias{resid,ANY-method}
\title{Method resid for package calib}
\description{
Methods for function \code{resid} in Package `calib'
}
\section{Methods}{
\describe{

\item{object = "ANY"}{Objects other than \code{calib.fit} for which \code{resid} can be used.}

\item{object = "calib.fit"}{\code{calib.fit} object.}
}}
\keyword{methods}
