\name{residuals-methods}
\docType{methods}
\alias{residuals}
\title{Residuals Method}
\description{
 Extract residual values from an object of class calib.fit
}
\section{Methods}{
\describe{

\item{object = "ANY"}{Extract residual values}

\item{object = "calib.fit"}{Extract residual values from a calib.fit object}
}}

\seealso{\code{\link{calib.fit}}}}

\keyword{methods}

