\name{RIAy0}
\alias{RIAy0}
\docType{data}
\title{Radio Immunoassay}
\description{
Dataset from the Davidian et. al 1996 paper.
}
\usage{data(RIAy0)}
\format{
  The format is:
 chr "RIAy0"
}

\source{
Dataset from the Davidian et al. 1996 paper.
}
\references{
Davidian et al. 1996
}
\examples{
data(RIAy0)
}
\keyword{datasets}
