\name{coefficients-methods}
\docType{methods}
\alias{coefficients-methods}
\title{Method coefficients for calib package}
\description{
Returns coefficient information from a \code{calib.fit} object.
}
\section{Methods}{
\describe{

\item{object = "ANY"}{Objects \code{coefficients} applies to other than \code{calib.fit} objects.}

\item{object = "calib.fit"}{\code{calib.fit} object.}
}}
\keyword{methods}
