\name{fitted-methods}
\docType{methods}
\alias{fitted-methods}
\alias{fitted,ANY-method}
\title{Method fitted for calib.fit object.}
\description{Returns fitted values from a \code{calib.fit} object.}
\section{Methods}{
\describe{

\item{object = "ANY"}{Objects other than \code{calib.fit} for which \code{fitted} can be used.}

\item{object = "calib.fit"}{\code{calib.fit} object.}
}}
\keyword{methods}
