\name{rdl-methods}
\docType{methods}
\alias{rdl}

\title{Reliable Detection Limit (RDL)}
\description{
 Extracts RDL value from an object of class calib.fit
}
\section{Methods}{
\describe{

\item{object = "calib.fit"}{Extracts RDL value}
}}
\seealso{\code{\link{calib.fit}}}}

\keyword{methods}

