\name{resid-methods}
\docType{methods}
\alias{resid-methods}
\alias{resid,ANY-method}
\title{Method resid for package calib}
\description{
Methods for function \code{resid} in Package `calib'
}
\section{Methods}{
\describe{

\item{object = "ANY"}{ ~~describe this method here }

\item{object = "calib.fit"}{ ~~describe this method here }
}}
\keyword{methods}
